243: move tests
